backend/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.25.2
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
python-dateutil==2.8.2
motor==3.4.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
aioredis==2.0.1
sqlalchemy==2.0.23
alembic==1.13.1